Extension does not follow redirections anymore.
General problem is the post/redirect/get pattern that prevent a post
request on the endpoint url of a redirection. There are some workarounds,
e.g. follow the redirection with get or head requests. But different servers
behave different. So there is at least no simple solution. Now users have
to add the exact marker url.